# labassist-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API สำหรับระบบจัดการ TA ภาควิชา CS มหาวิทยาลัยศิลปากร",
        "title": "LabAssist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "ดู activity log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "กรองตาม user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "GET",
                            "POST",
                            "PUT",
                            "DELETE"
                        ],
                        "type": "string",
                        "description": "กรองตาม HTTP method",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "หน้า (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "จำนวนต่อหน้า (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LogsResponse"
                        }
                    }
                }
            }
        },
        "/admin/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "สถิติภาพรวมของระบบ",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AdminStatsResponse"
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "รายการผู้ใช้ทั้งหมด",
                "parameters": [
                    {
                        "enum": [
                            "student",
                            "instructor",
                            "staff",
                            "admin"
                        ],
                        "type": "string",
                        "description": "กรองตาม role",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค้นหาด้วยชื่อหรืออีเมล",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "จำนวนต่อหน้า (default 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ออฟเซ็ต (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "สร้างผู้ใช้ใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลผู้ใช้",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "อัพเดตสถานะผู้ใช้ (เปิด/ปิด)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "สถานะใหม่",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateUserStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AdminStatsResponse": {
            "type": "object",
            "properties": {
                "accepted_applications": {
                    "type": "integer",
                    "example": 50
                },
                "open_courses": {
                    "type": "integer",
                    "example": 5
                },
                "pending_applications": {
                    "type": "integer",
                    "example": 30
                },
                "total_applications": {
                    "type": "integer",
                    "example": 200
                },
                "total_courses": {
                    "type": "integer",
                    "example": 15
                },
                "total_instructors": {
                    "type": "integer",
                    "example": 10
                },
                "total_students": {
                    "type": "integer",
                    "example": 80
                },
                "total_users": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "handlers.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "full_name",
                "password",
                "role",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@silpakorn.edu"
                },
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "example": "securepassword"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserRole"
                        }
                    ],
                    "example": "instructor"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "handlers.LogsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ActivityLog"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 50
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "handlers.UpdateUserStatusRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.ActivityLog": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "faculty": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "google_sub": {
                    "type": "string"
                },
                "gpa": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "student_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "student",
                "instructor",
                "staff",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleStudent",
                "RoleInstructor",
                "RoleStaff",
                "RoleAdmin"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}